Initial Commit on Isdale
To get this far on my mac, i had to archive the Newtonsoft.Json stuff under the QuickSheet as it conflicted with same stuff in the Mixed Reality Toolkit.
Still wont compile completely as something believes an MRTK file should be found in c:\Program Files (x86) on this mac.
silly tools.